Refactor Sort and Filter toolbar into its own component.
