+++ react-components/src/apps/model/viewType.js
@@ -1,0 +1,6 @@
+const viewType = {
+    TILE: "tile",
+    TABLE: "table",
+};
+
+export default viewType;
